requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
Werkzeug==1.0.1
